APP-24: Address PR feedback — require EXPO_PUBLIC_API_BASE_URL, drop 204 path, split types and endpoints into per-resource folders.
- Drop the hardcoded `DEFAULT_BASE_URL` fallback in `apiFetch`'s base URL resolver; throw when `EXPO_PUBLIC_API_BASE_URL` is unset.
- Drop the 204 No Content special case so `apiFetch<T>` always returns `Promise<T>` (no `T | undefined` leak).
- Split `app/api/types.ts` into per-resource files under `app/api/types/`.
- Split `app/api/endpoints/index.ts` into per-resource folders under `app/api/endpoints/<resource>/`, each with its own `.test.ts`.
- Update hook imports to point at the new file layout.
- Extract `jsonResponse` to a non-React `test-helpers.ts` so endpoint tests don't drag React in.
