+++ app/api/endpoints/replan/index.ts
@@ -1,0 +1,6 @@
+import { apiFetch } from '@/api/client';
+import type { ReplanResponse } from '@/api/types/replan';
+
+export function replan(): Promise<ReplanResponse> {
+    return apiFetch<ReplanResponse>('/replan', { method: 'POST' });
+}
