+++ app/api/endpoints/system/index.ts
@@ -1,0 +1,14 @@
+import { apiFetch } from '@/api/client';
+import type { SystemStateDto } from '@/api/types/system';
+
+export function getSystem(): Promise<SystemStateDto> {
+    return apiFetch<SystemStateDto>('/system');
+}
+
+export function enableSystem(): Promise<SystemStateDto> {
+    return apiFetch<SystemStateDto>('/system/enable', { method: 'POST' });
+}
+
+export function disableSystem(): Promise<SystemStateDto> {
+    return apiFetch<SystemStateDto>('/system/disable', { method: 'POST' });
+}
