+++ app/api/endpoints/tonight/index.ts
@@ -1,0 +1,6 @@
+import { apiFetch } from '@/api/client';
+import type { TonightDto } from '@/api/types/tonight';
+
+export function getTonight(): Promise<TonightDto> {
+    return apiFetch<TonightDto>('/tonight');
+}
