+++ app/api/test-utils.tsx
@@ -2,6 +2,8 @@ import { QueryClient } from '@tanstack/react-query';
 import type { ComponentType, PropsWithChildren } from 'react';
 import { ApiProvider } from '@/api/provider';
 
+export { jsonResponse } from '@/api/test-helpers';
+
 /**
  * Builds a fresh `<ApiProvider>` wrapper paired with an isolated
  * `QueryClient` so each hook test gets a clean cache and no cross-test
@@ -22,15 +24,4 @@ export function buildApiWrapper(): {
         <ApiProvider client={client}>{children}</ApiProvider>
     );
     return { wrapper, client };
-}
-
-/**
- * Builds a `Response` with a JSON body and the right content-type header.
- * Use as the resolved value of a `global.fetch` mock.
- */
-export function jsonResponse(body: unknown, status = 200): Response {
-    return new Response(JSON.stringify(body), {
-        status,
-        headers: { 'content-type': 'application/json' },
-    });
 }
