+++ app/api/types/zones.ts
@@ -1,0 +1,24 @@
+/**
+ * Wire shape returned by `GET /zones`. The api wraps the array in
+ * `{ zones: ZoneSummary[] }` — the endpoint wrapper unwraps for consumers.
+ */
+export type ZoneSummary = {
+    id: string;
+    slug: string;
+    name: string;
+    isEnabled: boolean;
+    grassType: { name: string };
+    soilType: { name: string };
+    areaM2: number;
+    rootDepthM: number;
+    allowableDepletionFraction: number;
+    irrigationEfficiency: number;
+    microclimateFactor: number;
+    precipitationRateMmPerHr: number | null;
+    currentDepletionMm: number;
+    rawMm: number;
+    lastFiredAt: string | null;
+    lastAppliedMm: number | null;
+    homeAssistantEntityId: string | null;
+    patch: string;
+};
