+++ app/hooks/tonight/index.ts
@@ -1,8 +1,8 @@
 import { useQuery, type UseQueryResult } from '@tanstack/react-query';
 import type { ApiError } from '@/api/client';
-import { getTonight } from '@/api/endpoints';
+import { getTonight } from '@/api/endpoints/tonight';
 import { keys } from '@/api/query-keys';
-import type { TonightDto } from '@/api/types';
+import type { TonightDto } from '@/api/types/tonight';
 
 /**
  * Returns the next-run summary for the Home hero card and CycleStrip. The
